refactor : changed edit and delete from submit mode to RESTFUL mode and added validation and handeled all errors in backend and front end

+++ MVC model/model/product.js
@@ -4,21 +4,9 @@ let books = [];
 
 const dbPath = path.join(__dirname, "../data/db.json");
 
-// const fetchAllBooks = () => {
-//   return new Promise((resolve, reject) => {
-//     fs.readFile(path.join(dbPath), "utf-8", (err, data) => {
-//       if (!err && data) {
-//         resolve({ data: JSON.parse(data) });
-//       } else {
-//         reject({ error: "you have no daata", data: [] });
-//       }
-//     });
-//   });
-// };
 const fetchAllBooks = (callback) => {
   fs.readFile(path.join(dbPath), "utf-8", (err, data) => {
     if (!err && data) {
-      // console.log(data);
       const prods = JSON.parse(data);
       callback(prods);
     } else {
@@ -48,28 +36,44 @@ class Product {
       });
     });
   }
-  static delete(bookId, cb) {
-    fetchAllBooks((prods) => {
-      books = prods;
-
-      books = books.filter((item) => item.id !== +bookId);
-
-      fs.writeFile(dbPath, JSON.stringify(books), (err) => {
-        if (err) console.log(err);
-        cb();
+  static delete(bookId) {
+    return new Promise((resolve, reject) => {
+      fetchAllBooks((prods) => {
+        books = prods;
+        const isBookExist = books.find((item) => item.id === +bookId);
+        if (isBookExist) {
+          books = books.filter((item) => item.id !== +bookId);
+          fs.writeFile(dbPath, JSON.stringify(books), (err) => {
+            if (err) {
+              reject({ success: false, message: err.cause });
+            } else resolve({ message: "product deleted successfully" });
+          });
+        } else {
+          reject({ message: "this book doesn't exist" });
+        }
       });
     });
   }
-  static editBook(bookId, newData, cb) {
-    fetchAllBooks((allBooks) => {
-      const book = allBooks.find((book) => book.id == +bookId);
-      for (const element in book) {
-        book[element] = newData[element];
-      }
-      book.id = Number(bookId);
-      fs.writeFile(dbPath, JSON.stringify(allBooks), (err) => {
-        console.log(err);
-        cb();
+  static editBook(bookId, newData) {
+    return new Promise((resolve, reject) => {
+      fetchAllBooks((allBooks) => {
+        const book = allBooks.find((book) => book.id == +bookId);
+        if (book) {
+          for (const key in book) {
+            book[key] = newData[key];
+          }
+          book.id = Number(bookId);
+          fs.writeFile(dbPath, JSON.stringify(allBooks), (err) => {
+            if (err) {
+              console.log(err);
+              reject({ success: false, message: err.cause });
+            } else {
+              resolve({ message: "product deleted successfully" });
+            }
+          });
+        } else {
+          reject({ success: false, message: "this book doesn't exist" });
+        }
       });
     });
   }
